cleaned up file structure
