--- content/webpack.config.js
@@ -1,28 +1,0 @@
-'use strict';
-
-var webpack = require('webpack');
-
-module.exports = {
-  entry: './src/scripts/index.js',
-  output: {
-    path: __dirname,
-    filename: './src/bundle.js'
-  },
-  context: __dirname,
-  devtool: 'source-map',
-  resolve: {
-    extensions: ['', '.js', '.jsx']
-  },
-  module: {
-    loaders: [
-      {
-        test: /jsx?$/,
-        exclude: /(node_modules|bower_components)/,
-        loader: 'babel',
-        query: {
-          presets: ['react', 'es2015', 'stage-2']
-        }
-      }
-    ]
-  }
-};
