+++ options/components/Settings.js
