+++ options/index.html
@@ -5,14 +5,12 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
     <title>RCR Example</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap.min.css">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap-theme.min.css">
     <link rel="stylesheet" href="style.css">
 </head>
 <body>
 
 <div id="app"></div>
 
-<script src="bundle.js"></script>
+<script src="../build/options.js"></script>
 </body>
 </html>
