+++ options/index.js
