+++ popup/index.html
@@ -10,6 +10,6 @@
 
 <div id="app"></div>
 
-<script src="bundle.js"></script>
+<script src="../build/popup.js"></script>
 </body>
 </html>
