--- popup/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "popup",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "webpack -w"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "babel": "^6.5.2",
-    "babel-core": "^6.18.2",
-    "babel-loader": "^6.2.7",
-    "babel-plugin-lodash": "^3.2.10",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-2": "^6.18.0",
-    "chalk": "^1.1.3",
-    "friendly-errors-webpack-plugin": "^1.1.0",
-    "node-oauth": "^0.1.3",
-    "npm-install-webpack-plugin": "^4.0.4",
-    "progress-bar-webpack-plugin": "^1.9.0",
-    "webpack": "^1.13.3"
-  },
-  "dependencies": {
-    "axios": "^0.15.2",
-    "bluebird": "^3.4.6",
-    "css-loader": "^0.26.0",
-    "file-loader": "^0.9.0",
-    "lodash": "^4.17.2",
-    "node-sass": "^3.13.0",
-    "rc-progress": "^2.0.6",
-    "react": "^15.4.0",
-    "react-chrome-redux": "^1.2.0",
-    "react-dom": "^15.4.0",
-    "react-motion": "^0.4.5",
-    "react-progressbar.js": "^0.2.0",
-    "react-redux": "^4.4.6",
-    "redux": "^3.6.0",
-    "redux-actions": "^0.13.0",
-    "redux-async-initial-state": "^0.2.0",
-    "redux-logger": "^2.7.4",
-    "redux-thunk": "^2.1.0",
-    "sass-loader": "^4.0.2",
-    "semantic-ui-react": "^0.64.4",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.7"
-  }
-}
